update controller and database
