+++ app/database.py
@@ -3,11 +3,12 @@ from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy import Boolean, Column, Integer, String
 from sqlalchemy.orm import sessionmaker
 
-SQLALCHEMY_DATABASE_URL = "mysql+pymysql://root:root@localhost/db?host=localhost?port=3306"
+SQLALCHEMY_DATABASE_URL = "mysql+pymysql://root:root@localhost/inventory"
 
 
 engine = create_engine(
-    SQLALCHEMY_DATABASE_URL
+    SQLALCHEMY_DATABASE_URL,
+    connect_args= dict(host='localhost', port=3306)
 )
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
